+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-# 1.0.1
+# WE 1.56
+## 1.1.0
+Fixed scene transitions: the free camera now should autodisable itself when the game switches to a different scene.
+Added compatibility with [WE_Street's Custom Arena Camera Mod](https://thunderstore.io/c/wrestling-empire/p/WE_Street/Custom_Arena_Camera_Mod/)
+## 1.0.1
 Audio and performance fixes